Format powers of ten

Format a numerical vector as a character string with power of ten notation in mathematical form or engineering form.

format_power(
  x,
  sig_dig = 3,
  ...,
  engr_notation = TRUE,
  disable_exponent = c(0.1, 1000)
)

Arguments

x

Numerical vector to be formatted.

sig_dig

Scalar integer, number of significant digits, default 3.

...

Not used, force later arguments to be used by name.

engr_notation

Logical, default TRUE. Places decimal points such that exponents are multiples of three. If FALSE, result has one digit to the left of the decimal point.

disable_exponent

Numeric vector of length two with the lower and upper limits over which scientific notation is disabled, default c(0.1, 1000).

Value

A character vector with the following properties:

Math powers of ten notation substituted for computer notation.

Delimited with $...$ for output as a mathematical expression to R Markdown or Quarto Markdown document.

Details

In engineering notation, exponents are multiples of three. Compatible with many of the output formats of R Markdown and Quarto Markdown. Can be applied to the numerical columns of a data frame via lapply().

Examples

# Scalar value
format_power(101100)
#> [1] "$101\\times10^{3}$"

# Vector value
x <- c(1.2222e-6, 2.3333e-5, 3.4444e-4, 4.1111e-3, 5.2222e-2, 6.3333e-1,
       7.4444e+0, 8.1111e+1, 9.2222e+2, 1.3333e+3, 2.4444e+4, 3.1111e+5, 4.2222e+6)
format_power(x)
#>  [1] "$1.22\\times10^{-6}$" "$23.3\\times10^{-6}$" "$344\\times10^{-6}$" 
#>  [4] "$4.11\\times10^{-3}$" "$52.2\\times10^{-3}$" "$0.633$"             
#>  [7] "$7.44$"               "$81.1$"               "$922$"               
#> [10] "$1.33\\times10^{3}$"  "$24.4\\times10^{3}$"  "$311\\times10^{3}$"  
#> [13] "$4.22\\times10^{6}$" 

# Compare significant digits
format_power(x[1], 3)
#> [1] "$1.22\\times10^{-6}$"
format_power(x[1], 4)
#> [1] "$1.222\\times10^{-6}$"

# Compare notation type
format_power(x[3], engr_notation = TRUE)
#> [1] "$344\\times10^{-6}$"
format_power(x[3], engr_notation = FALSE)
#> [1] "$3.44\\times10^{-4}$"

# Compare disabled exponents limits
format_power(x[6], disable_exponent = c(0.1, 1000))
#> [1] "$0.633$"
format_power(x[6], disable_exponent = c(1, 1000))
#> [1] "$633\\times10^{-3}$"

# Apply to columns of a data frame (data.table syntax)
y <- x[1:6]
z <- x[8:13]
DT <- data.table::data.table(y, z)
DT[, lapply(.SD, function(x) format_power(x))]
#>                       y                   z
#> 1: $1.22\\times10^{-6}$              $81.1$
#> 2: $23.3\\times10^{-6}$               $922$
#> 3:  $344\\times10^{-6}$ $1.33\\times10^{3}$
#> 4: $4.11\\times10^{-3}$ $24.4\\times10^{3}$
#> 5: $52.2\\times10^{-3}$  $311\\times10^{3}$
#> 6:              $0.633$ $4.22\\times10^{6}$