Format powers of ten

Convert the elements of a numerical vector to character strings in which the numbers are formatted using powers-of-ten notation in scientific or engineering form and delimited for rendering as inline equations in an R Markdown document.

Usage

format_power(x, digits = 3, ..., format = "engr", limits = c(0.1, 1000))

Arguments

x

Numeric vector to be formatted.

digits

Numeric scalar, nonzero positive integer to specify the number of significant digits in the output coefficient.

...

Not used, force later arguments to be used by name.

format

Character. Possible values are "engr" (engineering notation) and "sci" (scientific notation). Use by name.

limits

Numeric vector, length two, specifying the range over which power of ten notation is disabled and numbers are represented in decimal notation. Use by name.

Value

A character vector with the following properties:

Numbers within the limits range represented in decimal notation; all others represented in powers of ten notation.

Elements delimited with $...$ for rendering as inline math in an R Markdown or Quarto Markdown document.

Details

Given a number, a numerical vector, or a numerical column from a data frame, format_power() converts the numbers to character strings of the form, "$a\\times{10}^{n}$", where a is the coefficient and n is the exponent. The string includes markup delimiters $...$ for rendering the result as an inline equation in R Markdown or Quarto Markdown document. The user can specify the number of significant digits, scientific or engineering format, and the range over which decimal notation is enforced.

Examples

# Scalar value
format_power(101100)
#> [1] "$101\\times{10}^{3}$"

# Vector value
x <- c(1.2222e-6, 2.3333e-5, 3.4444e-4, 4.1111e-3, 5.2222e-2, 6.3333e-1,
       7.4444e+0, 8.1111e+1, 9.2222e+2, 1.3333e+3, 2.4444e+4, 3.1111e+5, 4.2222e+6)
format_power(x)
#>  [1] "$1.22\\times{10}^{-6}$" "$23.3\\times{10}^{-6}$" "$344\\times{10}^{-6}$" 
#>  [4] "$4.11\\times{10}^{-3}$" "$52.2\\times{10}^{-3}$" "$0.633$"               
#>  [7] "$7.44$"                 "$81.1$"                 "$922$"                 
#> [10] "$1.33\\times{10}^{3}$"  "$24.4\\times{10}^{3}$"  "$311\\times{10}^{3}$"  
#> [13] "$4.22\\times{10}^{6}$" 

# Compare significant digits
format_power(x[1], 3)
#> [1] "$1.22\\times{10}^{-6}$"
format_power(x[1], 4)
#> [1] "$1.222\\times{10}^{-6}$"

# Compare format type
format_power(x[3], format = "engr")
#> [1] "$344\\times{10}^{-6}$"
format_power(x[3], format = "sci")
#> [1] "$3.44\\times{10}^{-4}$"

# Compare change to limits
format_power(x[6], limits = c(0.1, 1000))
#> [1] "$0.633$"
format_power(x[6], limits = c(1, 1000))
#> [1] "$633\\times{10}^{-3}$"

# Apply to columns of a data frame (data.table syntax)
y <- x[1:6]
z <- x[8:13]
DT <- data.table::data.table(y, z)
DT[, lapply(.SD, function(x) format_power(x))]
#>                         y                     z
#> 1: $1.22\\times{10}^{-6}$                $81.1$
#> 2: $23.3\\times{10}^{-6}$                 $922$
#> 3:  $344\\times{10}^{-6}$ $1.33\\times{10}^{3}$
#> 4: $4.11\\times{10}^{-3}$ $24.4\\times{10}^{3}$
#> 5: $52.2\\times{10}^{-3}$  $311\\times{10}^{3}$
#> 6:                $0.633$ $4.22\\times{10}^{6}$