Format text to match the math font

Format a vector as character strings within math delimiters for rendering in an R Markdown document. The formatting is particularly useful for formatting text columns in a table in which numerical columns have been formatted using format_power() or format_decimal().

Usage

format_text(x, ..., face = NULL, size = NULL, delim = "$")

Arguments

x

Vector to be formatted.

...

Not used, force later arguments to be used by name.

face

Font face. Possible values are "plain", "italic", "bold", "sans", or "mono". Can also be set as a global option, for example, options(formatdown.font.face = "sans") that can be overwritten in an individual function call.

size

Font size. Possible values are "scriptsize", "small" (default), "normalsize", "large", and "huge". which correspond to selected LaTeX font size values. Can also be set as a global option, for example, options(formatdown.font.size = "normalsize") that can be overwritten in an individual function call.

delim

Character vector (length 1 or 2) defining the delimiters for marking up inline math. Possible values include "$" or "\\(", both of which create appropriate left and right delimiters. Alternatively, left and right can be defined explicitly in a character vector of length two, e.g., c("$", "$") or c("\\(", "\\)"). Custom delimiters can be assigned to suit the markup environment. Use argument by name.

Value

A character vector with elements delimited as inline math markup in plain, italic, or bold font face.

Details

Given a scalar, vector, or data frame column, format_text() converts its argument to a character string of the form "$\\mathxx{a}$" where a is the element to be formatted and \\mathxx determines the font face: plain type is set by \\mathrm; italic by \\mathit; bold by \\mathbf; sans serif by \\mathsf; and monospace (typewriter text) by \\mathtt. The string includes markup delimiters $...$ for rendering as an inline equation in R Markdown or Quarto Markdown document.

Using options() to assign a value to formatdown.font.face affects only that text formatted using format_text(). However, using options to assign formatdown.font.size affects output formatted by format_text() and format_power().

Delimiters for inline math markup can be edited if necessary. If the default argument fails, try using "\\(" as an alternative. If using a custom delimiter to suit the markup environment, be sure to escape all special symbols.

See also

Examples

# Text vector

# default face = "plain"
x <- air_meas$humid
format_text(x)
#> [1] "$\\small\\mathrm{low}$"  "$\\small\\mathrm{high}$"
#> [3] "$\\small\\mathrm{med}$"  "$\\small\\mathrm{low}$" 
#> [5] "$\\small\\mathrm{high}$"

# equivalently
format_text(x, face = "plain")
#> [1] "$\\small\\mathrm{low}$"  "$\\small\\mathrm{high}$"
#> [3] "$\\small\\mathrm{med}$"  "$\\small\\mathrm{low}$" 
#> [5] "$\\small\\mathrm{high}$"

# argument coerced to character string if possible
format_text(x = NA)
#> [1] "$\\small\\mathrm{NA}$"
format_text(x = c(TRUE, FALSE, TRUE))
#> [1] "$\\small\\mathrm{TRUE}$"  "$\\small\\mathrm{FALSE}$"
#> [3] "$\\small\\mathrm{TRUE}$" 

# numbers as strings are rendered as-is
format_text(x = c("1.2E-3", "3.4E+0", "5.6E+3"))
#> [1] "$\\small\\mathrm{1.2E-3}$" "$\\small\\mathrm{3.4E+0}$"
#> [3] "$\\small\\mathrm{5.6E+3}$"

# other font faces
format_text(x, face = "italic")
#> [1] "$\\small\\mathit{low}$"  "$\\small\\mathit{high}$"
#> [3] "$\\small\\mathit{med}$"  "$\\small\\mathit{low}$" 
#> [5] "$\\small\\mathit{high}$"
format_text(x, face = "bold")
#> [1] "$\\small\\mathbf{low}$"  "$\\small\\mathbf{high}$"
#> [3] "$\\small\\mathbf{med}$"  "$\\small\\mathbf{low}$" 
#> [5] "$\\small\\mathbf{high}$"
format_text(x, face = "sans")
#> [1] "$\\small\\mathsf{low}$"  "$\\small\\mathsf{high}$"
#> [3] "$\\small\\mathsf{med}$"  "$\\small\\mathsf{low}$" 
#> [5] "$\\small\\mathsf{high}$"
format_text(x, face = "mono")
#> [1] "$\\small\\mathtt{low}$"  "$\\small\\mathtt{high}$"
#> [3] "$\\small\\mathtt{med}$"  "$\\small\\mathtt{low}$" 
#> [5] "$\\small\\mathtt{high}$"

# alternate delimiter
format_text(x, delim = "\\(")
#> [1] "\\(\\small\\mathrm{low}\\)"  "\\(\\small\\mathrm{high}\\)"
#> [3] "\\(\\small\\mathrm{med}\\)"  "\\(\\small\\mathrm{low}\\)" 
#> [5] "\\(\\small\\mathrm{high}\\)"